requirements.txt
Django==2.2
django-crispy-forms==1.12.0
django-mathfilters==1.0.0
flake8==3.9.2
mccabe==0.6.1
Pillow==8.2.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
sqlparse==0.4.1
